Finalisation of instalScript new version with update capabilities

+++ instal/.instalModel/FILESinstal.sh
@@ -4,8 +4,8 @@
 #######################################################################
 # gestion des fichiers sources et copie dans les bons répertoires
 #######################################################################
-if [ -d /var/www/html ] ; then
-	if [ ! -d /var/www/html/public/cameraShots ] || [ ! -e /var/www/nichoir.db] ; then
+if [ -d /var/www/html ] && [ ! `ls -A /var/www/html | wc -c` -eq 0 ] ; then
+	if [ ! -d /var/www/html/public/cameraShots ] || [ ! -e /var/www/nichoir.db ] ; then
 		printMessage "Nettoyage du répertoire html" "rm -r /var/www/html/*"
 		rm -r /var/www/html/* >> $varLogFile 2>&1
 		printError "$?"
@@ -19,34 +19,36 @@ if [ "$varGit" == "true" ] ; then
 
 	# vérifie que les répertoires photo et film sont bien absent des fichiers
 	# avant de copier la nouvelle copie des fichiers sources
-	if [ ! -d eBirds/html_working/public/cameraShots ] ; then
-		rm -r -d eBirds/html_working/public/cameraShots 2>&1
+	if [ -d "eBirds/html_working/public/cameraShots" ] ; then
+		rm -r -d "eBirds/html_working/public/cameraShots" 2>&1
 	fi
 
-	if [ ! -d eBirds/html_working/public/cameraFilms ] ; then
-		rm -r -d eBirds/html_working/public/cameraFilms 2>&1
+	if [ -d "eBirds/html_working/public/cameraFilms" ] ; then
+		rm -r -d "eBirds/html_working/public/cameraFilms" 2>&1
 	fi
 
 	printMessage "déplacement des fichiers web" "/var/www/html"
-	mv --force eBirds/html_working/* /var/www/html/ >> $varLogFile 2>&1
+	sudo cp -r --force eBirds/html_working/* /var/www/html/ >> $varLogFile 2>&1
 	printError "$?"
 
 	# Crée des répertoires vide si non existant
-	if [ ! -d /var/www/html/public/cameraShots ] ; then
-		mkdir /var/www/html/public/cameraShots 2>&1
+	if [ ! -d "/var/www/html/public/cameraShots" ] ; then
+		sudo mkdir /var/www/html/public/cameraShots >> "$varLogFile" 2>&1
 	fi
 
-	if [ ! -d /var/www/html/public/cameraFilms ] ; then
-		mkdir /var/www/html/public/cameraFilms 2>&1
+	if [ ! -d "/var/www/html/public/cameraFilms" ] ; then
+		sudo mkdir "/var/www/html/public/cameraFilms" >> "$varLogFile" 2>&1
 	fi
 
 	printMessage "déplacement des scripts python" "/var/www/backend"
-	mv --force eBirds/backend /var/www/ >> $varLogFile 2>&1
+	sudo cp -r --force eBirds/backend /var/www/ >> "$varLogFile" 2>&1
 	printError "$?"
 
-	printMessage "déplacement du répertoire log" "/var/www/log"
-	mv --force eBirds/log /var/www/ >> $varLogFile 2>&1
-	printError "$?"
+	if [ ! -d "/var/www/log" ] ; then
+		printMessage "déplacement du répertoire log" "/var/www/log"
+		sudo mv --force eBirds/log /var/www/ >> "$varLogFile" 2>&1
+		printError "$?"
+	fi
 
 	printMessage "nettoyage des fichiers résiduels" "rm -r eBirds"
 	rm -r eBirds
